requirements.txt
astroid==2.1.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-extensions==2.1.4
django-markdownx==2.0.28
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
oauthlib==3.0.1
pickle-mixin==1.0.2
Pillow==5.4.1

protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pydotplus==2.0.2
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
schedule==0.5.0
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
sklearn==0.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.11.0
absl-py==0.7.0
astor==0.7.1
astroid==2.1.0
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
Django==2.1.7
django-extensions==2.1.4
django-markdownx==2.0.28
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
gast==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
grpcio==1.18.0
h5py==2.9.0
idna==2.8
isort==4.3.4
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
matplotlib==3.0.2
mccabe==0.6.1
numpy==1.16.1
oauthlib==3.0.1
pandas==0.24.1
pickle-mixin==1.0.2
Pillow==5.4.1

protobuf==3.6.1
psycopg2==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pydotplus==2.0.2
pylint==2.2.2
pylint-django==2.0.4
pylint-plugin-utils==0.4
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
schedule==0.5.0
scikit-learn==0.20.1
scipy==1.1.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
typed-ast==1.1.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.0
